libs/qec/pyproject.toml
[build-system]
requires = ["scikit-build-core>=0.10"]
build-backend = "scikit_build_core.build"

[project]
name = "cudaq-qec"
version = "0.1.0"
description = "Accelerated libraries for Quantum Error Correction built on CUDA-Q"
authors = [{name = "NVIDIA Corporation & Affiliates"}]
maintainers = [{name = "NVIDIA Corporation & Affiliates"}]
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
  'cuda-quantum-cu12 ~= 0.9.0',
]
classifiers = [
    'Intended Audience :: Science/Research',
    'Intended Audience :: Developers',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    "Environment :: GPU :: NVIDIA CUDA",
    "Environment :: GPU :: NVIDIA CUDA :: 12",
    'Topic :: Software Development',
    'Topic :: Scientific/Engineering',
]

[project.urls]
Homepage = "https://nvidia.github.io/cudaqx"
Documentation = "https://nvidia.github.io/cudaqx/components/qec/introduction.html"
Repository = "https://github.com/NVIDIA/cudaqx"

[tool.scikit-build]
build-dir = "_skbuild"
build.verbose = true
cmake.version = ">=3.28"
cmake.build-type = "Release"
install.components = ["qec-python", "qec-lib"]
wheel.packages = []
logging.level = "DEBUG"
ninja.version = ">=1.10"

[tool.scikit-build.cmake.define]
CUDAQX_QEC_INCLUDE_TESTS = false
CUDAQX_QEC_BINDINGS_PYTHON = true
